+++ src/biodwh2-pharmgkb/src/main/java/de/unibi/agbi/biodwh2/pharmgkb/etl/PharmGKBParser.java
@@ -1,29 +1,37 @@
 package de.unibi.agbi.biodwh2.pharmgkb.etl;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.univocity.parsers.common.processor.BeanListProcessor;
 import com.univocity.parsers.tsv.TsvParser;
 import com.univocity.parsers.tsv.TsvParserSettings;
 import de.unibi.agbi.biodwh2.core.Workspace;
 import de.unibi.agbi.biodwh2.core.etl.Parser;
 import de.unibi.agbi.biodwh2.core.exceptions.ParserException;
-import de.unibi.agbi.biodwh2.core.exceptions.ParserFileNotFoundException;
+import de.unibi.agbi.biodwh2.core.exceptions.ParserFormatException;
 import de.unibi.agbi.biodwh2.pharmgkb.PharmGKBDataSource;
 import de.unibi.agbi.biodwh2.pharmgkb.model.*;
+import de.unibi.agbi.biodwh2.pharmgkb.model.guideline.GuidelineAnnotation;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import java.io.*;
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
 import java.util.Enumeration;
 import java.util.List;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
 public class PharmGKBParser extends Parser<PharmGKBDataSource> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(PharmGKBParser.class);
+
     public PharmGKBParser(final PharmGKBDataSource dataSource) {
         super(dataSource);
     }
 
     @Override
     public boolean parse(final Workspace workspace) throws ParserException {
-        for (String filePath : PharmGKBUpdater.FILE_NAMES)
+        for (final String filePath : PharmGKBUpdater.FILE_NAMES)
             tryParseFile(workspace, dataSource, filePath);
         return true;
     }
@@ -33,20 +41,24 @@ public class PharmGKBParser extends Parser<PharmGKBDataSource> {
         try {
             parseFile(dataSource, dataSource.resolveSourceFilePath(workspace, filePath));
         } catch (IOException e) {
-            throw new ParserFileNotFoundException("Failed to parse the file '" + filePath + "'");
+            throw new ParserFormatException("Failed to parse the file '" + filePath + "'");
         }
     }
 
     private void parseFile(final PharmGKBDataSource dataSource, final String filePath) throws IOException {
-        ZipFile zipFile = new ZipFile(filePath);
-        Enumeration<? extends ZipEntry> entries = zipFile.entries();
+        final ZipFile zipFile = new ZipFile(filePath);
+        final Enumeration<? extends ZipEntry> entries = zipFile.entries();
         while (entries.hasMoreElements()) {
-            ZipEntry zipEntry = entries.nextElement();
-            String zipEntryName = zipEntry.getName();
-            InputStream stream = zipFile.getInputStream(zipEntry);
+            final ZipEntry zipEntry = entries.nextElement();
+            final String zipEntryName = zipEntry.getName();
+            final InputStream stream = zipFile.getInputStream(zipEntry);
             if (filePath.contains("pathways") && zipEntryName.startsWith("PA")) {
-                List<Pathway> tmpList = parseTSV(stream, Pathway.class);
+                final List<Pathway> tmpList = parseTSV(stream, Pathway.class);
                 dataSource.pathways.put(zipEntry.getName().split("\\.")[0], tmpList);
+            } else if (filePath.contains("dosingGuidelines.json.zip") && zipEntryName.endsWith(".json")) {
+                final GuidelineAnnotation annotation = parseGuidelineAnnotation(stream, zipEntryName);
+                if (annotation != null)
+                    dataSource.guidelineAnnotations.add(annotation);
             } else if (zipEntryName.equals("chemicals.tsv"))
                 dataSource.chemicals = parseTSV(stream, Chemical.class);
             else if (zipEntryName.equals("genes.tsv"))
@@ -86,13 +98,23 @@ public class PharmGKBParser extends Parser<PharmGKBDataSource> {
     }
 
     private <T> List<T> parseTSV(final InputStream stream, final Class<T> clazz) {
-        BeanListProcessor<T> processor = new BeanListProcessor<>(clazz);
-        TsvParserSettings parserSettings = new TsvParserSettings();
+        final BeanListProcessor<T> processor = new BeanListProcessor<>(clazz);
+        final TsvParserSettings parserSettings = new TsvParserSettings();
         parserSettings.setProcessor(processor);
         parserSettings.setHeaderExtractionEnabled(true);
         parserSettings.setMaxCharsPerColumn(200000);
-        TsvParser parser = new TsvParser(parserSettings);
-        parser.parse(stream, "UTF-8");
+        final TsvParser parser = new TsvParser(parserSettings);
+        parser.parse(stream, StandardCharsets.UTF_8);
         return processor.getBeans();
+    }
+
+    private GuidelineAnnotation parseGuidelineAnnotation(final InputStream stream, final String fileName) {
+        try {
+            final ObjectMapper objectMapper = new ObjectMapper();
+            return objectMapper.readValue(stream, GuidelineAnnotation.class);
+        } catch (IOException e) {
+            LOGGER.warn("Failed to parse guideline annotation '" + fileName + "'", e);
+        }
+        return null;
     }
 }